Added CRD for Topics and VIdeos

+++ utilities/helpers/topicHelpers.js
@@ -5,6 +5,7 @@ module.exports = {
   topicExists: async (name) => {
     try {
       const db = await getDatabase();
+      if (db instanceof AppError) return db;
       let exists = await db.execute(
         `SELECT COUNT(name) FROM topics WHERE name = '${name}' LIMIT 1`
       );
@@ -13,26 +14,56 @@ module.exports = {
       return new AppError(500, err.message);
     }
   },
-  insertTopic: async (topicName, topicDifficulty, topicDescription) => {
-    const db = await getDatabase();
-    await db.execute(
-      `INSERT INTO topics (name, description, difficulty) VALUES('${topicName}', '${topicDescription}', '${topicDifficulty}')`
-    );
-  },
-  getTopics: async () => {
+  insertTopic: async (topicName, topicDifficulty, topicDescription, username) => {
     try {
       const db = await getDatabase();
+      if (db instanceof AppError) return db;
+      await db.execute(
+        `INSERT INTO topics (name, description, difficulty, username) 
+        VALUES('${topicName}', '${topicDescription}', '${topicDifficulty}', '${username}')`
+      );
+    } catch (err) {
+      return new AppError(500, "Error creating topic");
+    }
+  },
+  getAllTopics: async () => {
+    try {
+      const db = await getDatabase();
+      if (db instanceof AppError) return db;
       let topics = await db.execute(
-        "SELECT name, description, difficulty FROM topics LIMIT 10"
+        "SELECT name, description, difficulty, username FROM topics LIMIT 10"
       );
       return topics[0].map((o) => Object.assign({}, o));
     } catch (err) {
       return new AppError(500, "Error Retrieving Topics");
     }
   },
+  getUserTopics: async (username) => {
+    try {
+      const db = await getDatabase();
+      if (db instanceof AppError) return db;
+      let topics = await db.execute(`SELECT name, description, difficulty 
+      FROM topics WHERE username = '${username}' LIMIT 10`);
+      return topics[0].map((o) => Object.assign({}, o));
+    } catch(err) {
+      return new AppError(500, `Error Retrieving ${username}'s topics`);
+    }
+  },
+  getTopic: async (topic) => {
+    try {
+      const db = await getDatabase();
+      if (db instanceof AppError) return db;
+      let topics = await db.execute(`SELECT name, description, difficulty 
+      FROM topics WHERE name = '${topic}' LIMIT 1`);
+      return topics[0].map((o) => Object.assign({}, o));
+    } catch (err) {
+      return new AppError(500, `Error Retrieving Topic`);
+    }
+  },
   removeTopic: async (topic) => {
     try {
       const db = await getDatabase();
+      if (db instanceof AppError) return db;
       await db.execute(`DELETE FROM topics WHERE name = '${topic}'`);
       return null;
     } catch (err) {
